Moving I/O decomp utils to spio_iodesc_utils.cpp
Moving some util functions related to I/O decomposition from pioc_support.cpp
to a new file (spio_iodesc_utils.cpp)

No change in source code
